requirements.txt
PyQt6>=6.0.0
websocket-client>=1.6.4
numpy>=1.26.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.1